refactor: Client-side socket event emission
- remove the redundant `handleRoom.js` and `handleNotification.js` modules.
- move `initSocial.js` to `socket` repository, and update the import path s in modules that utilize it.
- remove redundant `social` repository in `frontend/src/features/socket`.
